+++ src/api/index.js
@@ -1,3 +1,4 @@
+// @flow
 import Api from './Api';
 
 export { Api };