+++ CHANGELOG.md
@@ -2,7 +2,18 @@
 All notable changes to this project will be documented in this file.
 
 
-## [0.1.0] - 2018-04-04
+## [0.3.0] - 2018-04-07
+### Changed
+- Change method getActivityBetweenDates to getActivitiesBetweenDates
+- Add activity summary api
+
+### Fixed
+- Make user id a string
+- Fix retyping of subscription ids
+- Fixed missing flow descriptions
+
+
+## [0.2.0] - 2018-04-04
 ### Added
 - Add subscription api
 - Add activity summary api